fix #99. Return a separate time series
Previously we returned a single time series with the name "names" that
had one point per time series with the column "name" set to the time
series name. This patch changes the behavior of the server to return a
single time series with no points for each time series in the given db.