WIP: refactor so coordinator.RunQuery is the single entry point for running queries.
Coordinator will have the responsibility for distributing the query and possibly running through the engine. This is because the results returned by remote servers may not need to be run through the engine. I'll fix the engine tests later when this refactoring is completed to work with the new sharding style of clustering.