Wire up queries that require aggregation across shards.
Update coordinator to either pass through if aggregating locally or creating its own engine if it is required to aggregate.
Update shard logic to use passthrough aggregator if aggregation must be split across multiple shards.
Add method to querySpec that answers if the query will be split across multiple shards. This needs to be refined so it has better logic to handle queries that have multiple series, but can still be aggregated locally.