make sure that a shard endtime is exclusive when we search for shards range
note: this is hard to test since the shards will only return points in
the given range, so even if we use more shards than we're supposed to
the returned points will not be outside the query range. The only way to
test this is to write a unit test.