fix few bugs introduced in dbe76df
The changes in that commit broke Join and Merge, because they were
bypassing the QueryEngine which kept track of the series that should be
closed at the end of the query. Aggregators require an empty time series
to be sent after the query ends to flush their internal state.

The other bug was in deletes. Fixing the first bug broke the new
behavior of deletes which is return nothing if the time series is
empty. This is also fixed.