fix #342. Make sure we create a checkpoint before deleting data
Deleting data without creating a wal bookmark will cause the log to
replay some requests that where written before the delete happened,
causing the data to be resurrected. This change makes sure we create a
bookmark right before we delete the data to make sure that any write
before the delete will not be replayed.