more fixes for issue #341
We shouldn't be dropping responses anymore since the out of order
response reception isn't possible. Also fix the logic that decides
whether the shards should be queried sequentially or not. The only safe
case to do parallel querying is when we have a single time series with
aggregation over time only. Any other case is currently not safe to run
in parallel.