Fix #544. Add a way to force remove a node
This patch adds an endpoint as well as interfaces to RaftServer and
ClusterConfiguration to make it possible to forcibly remove a node from
the cluster. The way it works is similar to changing the connection
string. An InfluxJoinCommand command is created applied to local node as
well as other nodes in the cluster directly, then the command runs
through raft to make sure the change is permanent.