Fix some bugs with retention policy of shard spaces
This commit fixes two bugs:

Don't try to parse "inf" retention policy when creating a shard
space. This caused a panic to be thrown when a shard space is created
with infinity. Fix #774

`getExpiredShards()` used shard duration to determine which shards are
expired but should be using shard retention duration instead. Close #769