Fix the build on go1.3
There was one remaining reference to an incomplete struct type from c
which is value_array. The error printed by the compiler is intermittent,
meaning it compiles sometimes and print the error sometimes. Also,
value_array isn't strictly an incomplete data type since it contains a
pointer to an unknown data type but it's size could be computed.