Fix minus operator without surrounding spaces
Close #768

This patch isn't backward compatible since older version allowed the use
of table names with '-' without enclosing them. With this change, '-'
characters are disallowed completely unless they were enclosed in double
quotes. The parser can be made more complicated and be more context
aware to distinguish the use of the '-' character depending on the query
being parser, but I think the approach in this patch is simpler and less
confusing from the user's point of view.