Move cluster config from raft to API.
Fixes #853. Previously, there was an unprotected endpoint in raft to return the cluster config that would include user hashes. This endpoint is useful for debugging purposes so I restructured it and moved it to the API. It ensures the requesting user is a cluster admin.

Cluster config will now return all of the cluster state including servers, CQs, shards, etc.