Add server-level write batching.
This commit changes the Server to batch writes together -- even across
multiple requests. The server then periodically sends writes together
to the broker by shard in a single message.

This approach uses a lot of goroutines but it allows for all points to
manage the returned error and index. This simpler approach was used to
get basic batching working. I'll submit a separate PR for optimizing
batching to minimize goroutines.