Add simple write-and-query test case
The test case "TestHandler_serveWriteSeriesZeroTime" is actually failing
under certain circumstances, giving the impression that there is
something special about the case where the timestamp is not supplied. In
fact this is not the case, and even a write with a timestamp specified
is failing. Therefore add another test case that covers that case.