Add failing test for WHERE time() > now() - x
Currently, the test seems to fail based on the system time - e.g. if you
run the test at 5:32, the assertions pass for
`SELECT value FROM my_measurement WHERE time > now() - 60m` through
`... now() - 33m` but it will fail at `... now() - 32m`. If you ran this
at 1:55, it would pass for 60m-56m, but fail at 55m. The test passes if
you run it exactly on the hour.

Also noteworthy of this commit is that we are continuing to build on
other (currently) failing integration tests as discussed in team
standup.