Don't send broker-only messages to Replicas
Because these messages were being sent to replicas (data nodes) data
nodes were sometimes receiving messages out-of-order from the broker,
where order is defined strictly by the Raft log index. This causes our
tests to fail, because the new "wait" endpoint on broker assumes this
can never happen. When told to wait for, say, index 23, but sees it is
at 25, it responds with 200. But the message with ID 23 has yet to
arrive.

This change also means that if unknown messages are seen by data nodes,
they will instead panic, which should flag this issue much sooner.

Fixes issue #1556