Fix issue with time buckets and improve performance
When the start time fell in the middle of a group by bucket, the times for all buckets would be off by the amount of that stat time. This fixes it so group by windows are always on the bucket. However, if the user specifies a start time in the middle of a bucket, the very first bucket will only have whatever data in it that is past the start time (so it'll be a partial)

Also fixed a performance bug where if someone issued a query like: `SELECT mean(value) FROM foo GROUP BY time(10s) LIMIT 3`. Previously, the engine would create emtpy values for every 10s window since 1970, then limit it afterwards. Now, it'll just create value holders for the 3 buckets we need.