Data node fetches shard from peer when instructed
When a broker responds to a topic-stream request with list of data node
URLs, the requesting data node should instead fetch the shard for that
topic from one the URLs. When this fetch completes the data node writes
the shard to disk, and then restarts streaming from the broker at a
later index, which should be serviceable by the broker.