the development..
http://www.fsf.org/blogs/licensing/friday-free-software-directory-irc-meetup-april-10
</link>
<description>
Join the FSF and friends every Friday to help improve the Free Software Directory by adding new entries and updating existing ones.
</description>
<content:encoded xmlns:content="http://purl.org/rss/1.0/modules/content/">
<![CDATA[
<p>Join the FSF and friends today, Friday, April 10, from 2pm to 5pm EDT (18:00 to 21:00 UTC) to help improve the <a href="https://directory.fsf.org">Free Software Directory</a> by adding new entries and updating existing ones. We will be on IRC in the <a href="irc://irc.gnu.org/fsf">#fsf channel on freenode</a>.<br /> </p> <p>Tens of thousands of people visit <a href="https://directory.fsf.org">directory.fsf.org</a> each month to discover free software. Each entry in the Directory contains a wealth of useful information, from basic category and descriptions, to providing detailed info about version control, IRC channels, documentation, and licensing info that has been carefully checked by FSF staff and trained volunteers.<br /> </p> <p>While the Free Software Directory has been and continues to be a great resource to the world over the past decade, it has the potential of being a resource of even greater value. But it needs your help!<br /> </p> <p>If you are eager to help and you can't wait or are simply unable to make it onto IRC on Friday, our <a href="https://directory.fsf.org/wiki/FSD:Participate">participation guide</a> will provide you with all the information you need to get started on helping the Directory today!<br /> </p>
]]>
</content:encoded>
<dc:publisher>No publisher</dc:publisher>
<dc:creator/>
<dc:rights/>
<dc:date>2015-04-10T14:30:00Z</dc:date>
<dc:type>Blog Entry</dc:type>
</item>
<item rdf:about="http://www.fsf.org/blogs/licensing/obama-and-congress-go-off-the-rails-trying-to-fast-track-tpp">
<title>
Obama and Congress go off the rails trying to fast track TPP
</title>
<link>
http://www.fsf.org/blogs/licensing/obama-and-congress-go-off-the-rails-trying-to-fast-track-tpp
</link>
<description>
Fast tracking the TPP negotiations is back on the table. We have to fight back.
</description>
<content:encoded xmlns:content="http://purl.org/rss/1.0/modules/content/">
<![CDATA[
<p>While we are in the midst of the current <a href="https://www.defectivebydesign.org/end-DMCA-anti-circumvention-provisions">battle to end the DMCA's anti-circumvention provisions here in the US</a>, we can't lose sight of the broader global fight being waged via the secretive Trans-Pacific Partne
]]>
<![CDATA[
rship (TPP) Strategic Partnership Agreement. If you aren't familiar with TPP, it is a multinational trade agreement being developed through a series of secret negotiations that are pushing a host of restrictions. From making the DMCA's anti-circumvention provisions global, to spreading the threat of software patents around the world, to extending copyright indefinitely, these secret negotiations present a plethora of threats to user freedom. </p> <p>Since our last <a href="https://www.fsf.org/blogs/licensing/the-sharks-move-in-lobbyists-pushing-forward-on-tpp-agreements">update on TPP</a>, the Obama administration and the Republican-led Congress have been maneuvering to put the TPP nightmare on a fast track. In essence, they want to ensure that whatever the outcome of the TPP negotiations, the TPP will be immediately etched into U.S. law without real debate. It is bad enough that while the negotiations are ongoing, we are forced to guess at their effects based only on leaked information. That Congress won't even be debating the outcome before implementing these terrible restrictions in US law moves this issue into the realm of the surreal. If TPP is fast-tracked, there will be no debate, no amendments, no chance to limit or remove the terrible things currently being debated in secret. </p> <p>We need to stop the Obama administration and his allies in the Republican-led Congress before they railroad us into a future run by secret lobbyists. A future where circumventing Digital Restrictions Management is a crime everywhere. A future where copyright extends beyond all right and reason in perpetuity. A future where software patents threaten the development of free software everywhere. We have to send a clear message to Congress and the administration, telling them that they must not fast track TPP. This is our stop: it's time to get off.</p> <p>This is what you can do to help derail TPP before it becomes a reality.</p> <ul> <li>If you are in the US, join EFF's action and <a href="https://act.eff.org-**********= = =<<<<<>>>,,