Fix panic: reflect: call of reflect.Value.NumField on string Value
The env var overrides panic if we if peers was set in the config.
The type of that value is a []string and NumFields is not a valid
for that type.  We don't currently support this slice fields via
the env variable settings.  This particular one can be set with the
-join flag already so we just skip these fields for now.